# Conflicts:
#	src/Controller/ReceptionController.php

+++ src/Controller/DemandeController.php
@@ -2,11 +2,10 @@
 
 namespace App\Controller;
 
+use App\Entity\Action;
 use App\Entity\Demande;
+use App\Entity\Menu;
 use App\Entity\Preparation;
-
-use App\Form\LivraisonType;
-
 use App\Repository\DemandeRepository;
 use App\Repository\ReferenceArticleRepository;
 use App\Repository\LigneArticleRepository;
@@ -14,6 +13,7 @@ use App\Repository\StatutRepository;
 use App\Repository\EmplacementRepository;
 use App\Repository\UtilisateurRepository;
 
+use App\Service\UserService;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
@@ -56,7 +56,13 @@ class DemandeController extends AbstractController
      */
     private $referenceArticleRepository;
 
-    public function __construct(LigneArticleRepository $ligneArticleRepository, DemandeRepository $demandeRepository, StatutRepository $statutRepository, ReferenceArticleRepository $referenceArticleRepository, UtilisateurRepository $utilisateurRepository, EmplacementRepository $emplacementRepository)
+    /**
+     * @var UserService
+     */
+    private $userService;
+
+
+    public function __construct(LigneArticleRepository $ligneArticleRepository, DemandeRepository $demandeRepository, StatutRepository $statutRepository, ReferenceArticleRepository $referenceArticleRepository, UtilisateurRepository $utilisateurRepository, EmplacementRepository $emplacementRepository, UserService $userService)
     {
         $this->statutRepository = $statutRepository;
         $this->emplacementRepository = $emplacementRepository;
@@ -64,6 +70,7 @@ class DemandeController extends AbstractController
         $this->utilisateurRepository = $utilisateurRepository;
         $this->referenceArticleRepository = $referenceArticleRepository;
         $this->ligneArticleRepository = $ligneArticleRepository;
+        $this->userService = $userService;
     }
 
     /**
@@ -72,6 +79,10 @@ class DemandeController extends AbstractController
     public function finish(Request $request): Response
     {
         if (!$request->isXmlHttpRequest() && $data = json_decode($request->getContent(), true)) {
+            if (!$this->userService->hasRightFunction(Menu::PREPA, Action::CREATE)) {
+                return $this->redirectToRoute('access_denied');
+            }
+
             // Creation d'une nouvelle preparation basée sur une selection de demandes
             $demande = $this->demandeRepository->find($data['demande']);
             $preparation = new Preparation();
@@ -98,23 +109,24 @@ class DemandeController extends AbstractController
                     'demande/enteteDemandeLivraison.html.twig',
                     [
                         'demande' => $demande,
+                        'modifiable' => ($demande->getStatut()->getNom() === (Demande::STATUT_BROUILLON)),
                     ]
                 ),
             ];
 
             return new JsonResponse($data);
         }
-        throw new NotFoundHttpException("404"); //TODO retour msg erreur (pas d'article dans la DL)
+        throw new NotFoundHttpException('404'); //TODO retour msg erreur (pas d'article dans la DL)
     }
 
-   /**
+    /**
      * @Route("/api-modifier", name="demandeLivraison_api_edit", options={"expose"=true}, methods="GET|POST")
      */
     public function editApi(Request $request): Response
     {
         if (!$request->isXmlHttpRequest() && $data = json_decode($request->getContent(), true)) {
             $demandeLivraison = $this->demandeRepository->find($data);
-           
+
             $json = $this->renderView('demande/modalEditDemandeContent.html.twig', [
                 'demande' => $demandeLivraison,
             ]);
@@ -130,8 +142,11 @@ class DemandeController extends AbstractController
     public function edit(Request $request): Response
     {
         if (!$request->isXmlHttpRequest() && $data = json_decode($request->getContent(), true)) {
-           
-            $utilisateur = $this->utilisateurRepository->find(intval($data["demandeur"]));
+            if (!$this->userService->hasRightFunction(Menu::DEM_LIVRAISON, Action::CREATE)) {
+                return $this->redirectToRoute('access_denied');
+            }
+
+            $utilisateur = $this->utilisateurRepository->find(intval($data['demandeur']));
             $emplacement = $this->emplacementRepository->find(intval($data['destination']));
             $demande = $this->demandeRepository->find($data['demandeId']);
             $demande
@@ -144,11 +159,13 @@ class DemandeController extends AbstractController
             $json = [
                 'entete' => $this->renderView('demande/enteteDemandeLivraison.html.twig', [
                     'demande' => $demande,
-                ])
+                    'modifiable' => ($demande->getStatut()->getNom() === (Demande::STATUT_BROUILLON)),
+                ]),
             ];
+
             return new JsonResponse($json);
         }
-        throw new NotFoundHttpException("404");
+        throw new NotFoundHttpException('404');
     }
 
     /**
@@ -157,11 +174,15 @@ class DemandeController extends AbstractController
     public function new(Request $request): Response
     {
         if (!$request->isXmlHttpRequest() && $data = json_decode($request->getContent(), true)) {
+            if (!$this->userService->hasRightFunction(Menu::DEM_LIVRAISON, Action::CREATE)) {
+                return $this->redirectToRoute('access_denied');
+            }
+
             $em = $this->getDoctrine()->getManager();
-            $utilisateur = $this->utilisateurRepository->find($data["demandeur"]);
+            $utilisateur = $this->utilisateurRepository->find($data['demandeur']);
             $date = new \DateTime('now');
             $statut = $this->statutRepository->findOneByCategorieAndStatut(Demande::CATEGORIE, Demande::STATUT_BROUILLON);
-            $destination = $this->emplacementRepository->find($data["destination"]);
+            $destination = $this->emplacementRepository->find($data['destination']);
             $demande = new Demande();
             $demande
                 ->setStatut($statut)
@@ -175,19 +196,23 @@ class DemandeController extends AbstractController
             $em->flush();
 
             $data = [
-                "redirect" => $this->generateUrl('ligne_article_show', ['id' => $demande->getId()])
+                "redirect" => $this->generateUrl('demande_show', ['id' => $demande->getId()])
             ];
 
             return new JsonResponse($data);
         }
-        throw new NotFoundHttpException("404");
+        throw new NotFoundHttpException('404');
     }
 
     /**
      * @Route("/", name="demande_index", methods={"GET"})
      */
-    public function index(Request $request): Response
+    public function index(): Response
     {
+        if (!$this->userService->hasRightFunction(Menu::DEM_LIVRAISON, Action::LIST)) {
+            return $this->redirectToRoute('access_denied');
+        }
+
         return $this->render('demande/index.html.twig', [
             'utilisateurs' => $this->utilisateurRepository->getIdAndUsername(),
             'statuts' => $this->statutRepository->findByCategorieName(Demande::CATEGORIE),
@@ -201,16 +226,21 @@ class DemandeController extends AbstractController
     public function delete(Request $request): Response
     {
         if (!$request->isXmlHttpRequest() && $data = json_decode($request->getContent(), true)) {
+            if (!$this->userService->hasRightFunction(Menu::DEM_LIVRAISON, Action::DELETE)) {
+                return $this->redirectToRoute('access_denied');
+            }
+
             $demande = $this->demandeRepository->find($data['demandeId']);
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($demande);
             $entityManager->flush();
             $data = [
-                "redirect" => $this->generateUrl('demande_index')
+                'redirect' => $this->generateUrl('demande_index'),
             ];
+
             return new JsonResponse($data);
         }
-        throw new NotFoundHttpException("404");
+        throw new NotFoundHttpException('404');
     }
 
     /**
@@ -219,13 +249,15 @@ class DemandeController extends AbstractController
     public function api(Request $request): Response
     {
         if ($request->isXmlHttpRequest()) {
+            if (!$this->userService->hasRightFunction(Menu::DEM_LIVRAISON, Action::LIST)) {
+                return $this->redirectToRoute('access_denied');
+            }
 
-            // $demandes = $demandeRepository->findByUserAndNotStatus($this->getUser(), Livraison::STATUT_TERMINE);
             $demandes = $this->demandeRepository->findAll();
             $rows = [];
             foreach ($demandes as $demande) {
                 $idDemande = $demande->getId();
-                $url = $this->generateUrl('ligne_article_show', ['id' => $idDemande]);
+                $url = $this->generateUrl('demande_show', ['id' => $idDemande]);
                 $rows[] =
                     [
                         "Date" => ($demande->getDate() ? $demande->getDate()->format('d/m/Y') : ''),
@@ -244,21 +276,160 @@ class DemandeController extends AbstractController
             $data['data'] = $rows;
             return new JsonResponse($data);
         }
+        throw new NotFoundHttpException('404');
+    }
+
+    /**
+     * @Route("/voir/{id}", name="demande_show", methods={"GET", "POST"})
+     */
+    public function show(Demande $demande): Response
+    {
+        if (!$this->userService->hasRightFunction(Menu::DEM_LIVRAISON, Action::LIST)) {
+            return $this->redirectToRoute('access_denied');
+        }
+
+        return $this->render('demande/show.html.twig', [
+            'demande' => $demande,
+            'utilisateurs' => $this->utilisateurRepository->getIdAndUsername(),
+            'statuts' => $this->statutRepository->findByCategorieName(Demande::CATEGORIE),
+            'references' => $this->referenceArticleRepository->getIdAndLibelle(),
+            'modifiable' => ($demande->getStatut()->getNom() === (Demande::STATUT_BROUILLON)),
+            'emplacements' => $this->emplacementRepository->findAll()
+        ]);
+    }
+
+    /**
+     * @Route("/api/{id}", name="demande_article_api", options={"expose"=true},  methods="GET|POST")
+     */
+    public function articleApi(Request $request, Demande $demande): Response
+    {
+        if ($request->isXmlHttpRequest()) {
+            if (!$this->userService->hasRightFunction(Menu::DEM_LIVRAISON, Action::LIST)) {
+                return $this->redirectToRoute('access_denied');
+            }
+
+            $ligneArticles = $demande->getLigneArticle();
+            $rows = [];
+            foreach ($ligneArticles as $ligneArticle) {
+                $idArticle = $ligneArticle->getId();
+                $url['delete'] = $this->generateUrl('demande_remove_article', ['id' => $ligneArticle->getId()]);
+                $rows[] = [
+                    "Référence CEA" => ($ligneArticle->getReference()->getReference() ? $ligneArticle->getReference()->getReference() : ''),
+                    "Libellé" => ($ligneArticle->getReference()->getLibelle() ? $ligneArticle->getReference()->getLibelle() : ''),
+                    "Quantité" => ($ligneArticle->getQuantite() ? $ligneArticle->getQuantite() : ''),
+                    "Actions" => $this->renderView(
+                        'demande/datatableLigneArticleRow.html.twig',
+                        [
+                            'idArticle' => $idArticle,
+                            'modifiable' => ($demande->getStatut()->getNom() === (Demande::STATUT_BROUILLON)),
+                        ]
+                    )
+                ];
+            }
+
+            $data['data'] = $rows;
+            return new JsonResponse($data);
+        }
         throw new NotFoundHttpException("404");
     }
 
     /**
-     * @Route("/voir", name="demande_show", options={"expose"=true}, methods={"GET", "POST"})
+     * @Route("/ajouter-article", name="demande_add_article", options={"expose"=true},  methods="GET|POST")
      */
-    public function show(Request $request): Response
+    public function addArticle(Request $request): Response
     {
         if (!$request->isXmlHttpRequest() && $data = json_decode($request->getContent(), true)) {
-            $demande = $this->demandeRepository->find($data);
-            $json = $this->renderView('demande/modalEditDemand.html.twig', [
-                'demande' => $demande,
+            if (!$this->userService->hasRightFunction(Menu::DEM_LIVRAISON, Action::CREATE)) {
+                return $this->redirectToRoute('access_denied');
+            }
+
+            $referenceArticle = $this->referenceArticleRepository->find($data["reference"]);
+            $demande = $this->demandeRepository->find($data['demande']);
+            if ($this->ligneArticleRepository->countByRefArticleDemande($referenceArticle, $demande) < 1) {
+                $ligneArticle = new LigneArticle();
+                $ligneArticle
+                    ->setQuantite($data["quantite"])
+                    ->setReference($referenceArticle);
+            } else {
+                $ligneArticle = $this->ligneArticleRepository->getByRefArticle($referenceArticle);
+                $ligneArticle
+                    ->setQuantite($ligneArticle->getQuantite() + $data["quantite"]);
+            }
+
+            $quantiteReservee = intval($data["quantite"]);
+            $quantiteArticleReservee = $referenceArticle->getQuantiteReservee();
+
+            $referenceArticle
+                ->setQuantiteReservee($quantiteReservee + $quantiteArticleReservee);
+            $demande
+                ->addLigneArticle($ligneArticle);
+
+            $em = $this->getDoctrine()->getEntityManager();
+            $em->persist($ligneArticle);
+            $em->flush();
+
+            return new JsonResponse();
+        }
+        throw new NotFoundHttpException("404");
+    }
+
+    /**
+     * @Route("/retirer-article", name="demande_remove_article", options={"expose"=true}, methods={"GET", "POST"})
+     */
+    public function removeArticle(Request $request): Response
+    {
+        if (!$request->isXmlHttpRequest() && $data = json_decode($request->getContent(), true)) {
+            if (!$this->userService->hasRightFunction(Menu::DEM_LIVRAISON, Action::CREATE)) {
+                return $this->redirectToRoute('access_denied');
+            }
+
+            $ligneAricle = $this->ligneArticleRepository->find($data['ligneArticle']);
+            $entityManager = $this->getDoctrine()->getManager();
+            $entityManager->remove($ligneAricle);
+            $entityManager->flush();
+            return new JsonResponse();
+        }
+        throw new NotFoundHttpException("404");
+    }
+
+    /**
+     * @Route("/modifier-article", name="demande_article_edit", options={"expose"=true}, methods={"GET", "POST"})
+     */
+    public function editArticle(Request $request): Response
+    {
+        if (!$request->isXmlHttpRequest() && $data = json_decode($request->getContent(), true)) {
+            if (!$this->userService->hasRightFunction(Menu::DEM_LIVRAISON, Action::CREATE)) {
+                return $this->redirectToRoute('access_denied');
+            }
+
+            $reference = $this->referenceArticleRepository->find($data['reference']);
+            $ligneArticle = $this->ligneArticleRepository->find($data['ligneArticle']);
+            $ligneArticle
+                ->setReference($reference)
+                ->setQuantite($data["quantite"]);
+            $this->getDoctrine()->getEntityManager()->flush();
+
+            return new JsonResponse();
+        }
+        throw new NotFoundHttpException("404");
+    }
+
+    /**
+     * @Route("/api-modifier", name="demande_article_api_edit", options={"expose"=true}, methods={"POST"})
+     */
+    public function articleEditApi(Request $request): Response
+    {
+        if (!$request->isXmlHttpRequest() && $data = json_decode($request->getContent(), true)) {
+            if (!$this->userService->hasRightFunction(Menu::DEM_LIVRAISON, Action::CREATE)) {
+                return $this->redirectToRoute('access_denied');
+            }
+
+            $ligneArticle = $this->ligneArticleRepository->getQuantity($data);
+            $json = $this->renderView('demande/modalEditArticleContent.html.twig', [
+                'ligneArticle' => $ligneArticle,
             ]);
             return new JsonResponse($json);
         }
-        throw new NotFoundHttpException("404");
+        throw new NotFoundHttpException('404');
     }
 }
